Fixing LoginController logic

+++ controller/LoginController.php
@@ -3,23 +3,19 @@
 namespace Controller;
 
 class LoginController extends Controller
-
 {
     private $isLoggedIn = false;
+    private $formAction;
+
     public function updateState(): void
     {
         try {
-            $form = $this->view->getForm();
-            $this->isLoggedIn = \Model\UserStorage::validateSession();
-            if (!$this->isLoggedIn) {
-                $this->loginWithCookies();
-            }
-            if (!$this->isLoggedIn && $form->getAction() == \Model\FormAction::$login) {
-                $this->attemptLogin();
-            }
-            if ($this->isLoggedIn && $form->getAction() == \Model\FormAction::$logout) {
-                $this->logout();
-            }
+            //TODO: rename these methods..
+            $this->formAction = $this->view->getFormAction();
+            $this->attemptLoginWithSession();
+            $this->attemptLoginWithCookies();
+            $this->attemptLoginWithLoginForm();
+            $this->attemptLogout();
             $this->view->setIsLoggedIn($this->isLoggedIn);
         } catch (\Exception $e) {
             $this->view->setMessage($e->getMessage());
@@ -36,8 +32,17 @@ class LoginController extends Controller
         $this->view->setMessage($message);
     }
 
-    private function loginWithCookies(): void
+    private function attemptLoginWithSession(): void
     {
+        $this->isLoggedIn = \Model\UserStorage::validateSession();
+    }
+
+    private function attemptLoginWithCookies(): void
+    {
+        if ($this->isLoggedIn) {
+            return;
+        }
+
         $cookies = $this->view->getCookies();
         if ($cookies) {
             $this->isLoggedIn = \Model\UserStorage::validateCookies($cookies);
@@ -46,28 +51,44 @@ class LoginController extends Controller
         }
     }
 
-    private function attemptLogin(): void
+    private function attemptLoginWithLoginForm(): void
     {
-        $form = $this->view->getForm();
-        $this->isLoggedIn = \Model\UserStorage::validateUserCredentials($form);
-        $this->view->setMessage($this->isLoggedIn ? \Model\Messages::$welcome : \Model\Messages::$incorrectCredentials);
-        $this->saveSession($form->getUsername());
-        $this->view->setCookies();
+        if (!$this->isLoggedIn && $this->formAction == \Model\FormAction::$login) {
+
+            $username = $this->view->getFormUsername();
+            $password = $this->view->getFormPassword();
+            $form = new \Model\LoginCredentials($username, $password);
+            $this->isLoggedIn = \Model\UserStorage::validateUserCredentials($form);
+            $this->view->setMessage(
+                $this->isLoggedIn
+                    ? \Model\Messages::$welcome
+                    : \Model\Messages::$incorrectCredentials
+            );
+            $this->saveSession($form->getUsername());
+            $this->setCookies();
+        }
     }
 
-    private function logout(): void
+    private function attemptLogout(): void
     {
-        $this->isLoggedIn = false;
-        \Model\UserStorage::destroySession();
-        $this->view->unsetCookies();
-        $this->view->setMessage(\Model\Messages::$logout);
+        if ($this->isLoggedIn && $this->formAction == \Model\FormAction::$logout) {
+            $this->isLoggedIn = false;
+            \Model\UserStorage::destroySession();
+            $this->view->unsetCookies();
+            $this->view->setMessage(\Model\Messages::$logout);
+        }
     }
-
 
     private function saveSession(string $username): void
     {
         if ($this->isLoggedIn) {
             \Model\UserStorage::saveSession($username);
         }
+    }
+
+    private function setCookies(): void
+    {
+        // TODO: control cookie login here. Ask view, tell view. No login in view.
+        $this->view->setCookies();
     }
 }
